+++ data-ingestion/__init__.py
@@ -1,2 +1,2 @@
-"""data-ingestion service v0.3.0"""
-__version__ = "0.3.0"
+"""data-ingestion service v0.4.0"""
+__version__ = "0.4.0"